Move `ValidationConfig` to `validation`
