+++ crates/fj-core/src/objects/any_object.rs
@@ -4,8 +4,8 @@ use crate::{
         Surface, Vertex,
     },
     storage::{Handle, HandleWrapper, ObjectId},
-    validate::{Validate, ValidationConfig},
-    validation::ValidationError,
+    validate::Validate,
+    validation::{ValidationConfig, ValidationError},
 };
 
 macro_rules! any_object {
