+++ crates/fj-core/src/validate/curve.rs
@@ -1,6 +1,9 @@
-use crate::{objects::Curve, validation::ValidationError};
+use crate::{
+    objects::Curve,
+    validation::{ValidationConfig, ValidationError},
+};
 
-use super::{Validate, ValidationConfig};
+use super::Validate;
 
 impl Validate for Curve {
     fn validate_with_config(
