+++ crates/fj-core/src/validate/mod.rs
@@ -73,7 +73,10 @@ mod solid;
 mod surface;
 mod vertex;
 
-use crate::{storage::ObjectId, validation::ValidationError};
+use crate::{
+    storage::ObjectId,
+    validation::{ValidationConfig, ValidationError},
+};
 
 pub use self::{
     cycle::CycleValidationError, edge::EdgeValidationError,
@@ -82,8 +85,6 @@ pub use self::{
 };
 
 use std::{collections::HashMap, error::Error, thread};
-
-use fj_math::Scalar;
 
 /// Assert that some object has a validation error which matches a specific
 /// pattern. This is preferred to matching on [`Validate::validate_and_return_first_error`], since usually we don't care about the order.
@@ -174,36 +175,4 @@ pub trait Validate: Sized {
         config: &ValidationConfig,
         errors: &mut Vec<ValidationError>,
     );
-}
-
-/// Configuration required for the validation process
-#[derive(Debug, Clone, Copy)]
-pub struct ValidationConfig {
-    /// The minimum distance between distinct objects
-    ///
-    /// Objects whose distance is less than the value defined in this field, are
-    /// considered identical.
-    pub distinct_min_distance: Scalar,
-
-    /// The maximum distance between identical objects
-    ///
-    /// Objects that are considered identical might still have a distance
-    /// between them, due to inaccuracies of the numerical representation. If
-    /// that distance is less than the one defined in this field, can not be
-    /// considered identical.
-    pub identical_max_distance: Scalar,
-}
-
-impl Default for ValidationConfig {
-    fn default() -> Self {
-        Self {
-            distinct_min_distance: Scalar::from_f64(5e-7), // 0.5 µm,
-
-            // This value was chosen pretty arbitrarily. Seems small enough to
-            // catch errors. If it turns out it's too small (because it produces
-            // false positives due to floating-point accuracy issues), we can
-            // adjust it.
-            identical_max_distance: Scalar::from_f64(5e-14),
-        }
-    }
 }
